Refactor: Moved ChatList to components
